requirements.txt
asgiref==3.6.0
Django==4.1.7
django-appconf==1.0.5
django-crontab==0.7.1
django-imagekit==4.1.0
djangorestframework==3.14.0
easy-thumbnails==2.8.5
pilkit==2.0
Pillow==9.4.0
python-magic==0.4.27
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3